feature: TRACEFOSS-1412 refactored hasPolicyContraint
