+++ tx-backend/src/main/java/org/eclipse/tractusx/traceability/infrastructure/edc/blackbox/EdcService.java
@@ -23,19 +23,13 @@ package org.eclipse.tractusx.traceability.infrastructure.edc.blackbox;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import okhttp3.Request;
 import okhttp3.RequestBody;
-import org.eclipse.tractusx.traceability.infrastructure.edc.blackbox.asset.Asset;
 import org.eclipse.tractusx.traceability.infrastructure.edc.blackbox.catalog.Catalog;
 import org.eclipse.tractusx.traceability.infrastructure.edc.blackbox.notification.ContractNegotiationDto;
 import org.eclipse.tractusx.traceability.infrastructure.edc.blackbox.notification.ContractOfferDescription;
 import org.eclipse.tractusx.traceability.infrastructure.edc.blackbox.notification.NegotiationInitiateRequestDto;
 import org.eclipse.tractusx.traceability.infrastructure.edc.blackbox.notification.TransferId;
 import org.eclipse.tractusx.traceability.infrastructure.edc.blackbox.offer.ContractOffer;
-import org.eclipse.tractusx.traceability.infrastructure.edc.blackbox.policy.AtomicConstraint;
-import org.eclipse.tractusx.traceability.infrastructure.edc.blackbox.policy.Expression;
-import org.eclipse.tractusx.traceability.infrastructure.edc.blackbox.policy.LiteralExpression;
-import org.eclipse.tractusx.traceability.infrastructure.edc.blackbox.policy.Permission;
 import org.eclipse.tractusx.traceability.infrastructure.edc.blackbox.policy.Policy;
-import org.eclipse.tractusx.traceability.infrastructure.edc.blackbox.policy.Rule;
 import org.eclipse.tractusx.traceability.infrastructure.edc.blackbox.transfer.DataAddress;
 import org.eclipse.tractusx.traceability.infrastructure.edc.blackbox.transfer.TransferRequestDto;
 import org.eclipse.tractusx.traceability.infrastructure.edc.blackbox.transfer.TransferType;
@@ -46,216 +40,146 @@ import org.springframework.stereotype.Component;
 
 import java.io.IOException;
 import java.lang.invoke.MethodHandles;
-import java.util.Collection;
-import java.util.List;
 import java.util.Map;
 import java.util.Optional;
-import java.util.concurrent.*;
+import java.util.concurrent.ExecutionException;
+import java.util.concurrent.Executors;
+import java.util.concurrent.ScheduledExecutorService;
+import java.util.concurrent.ScheduledFuture;
+import java.util.concurrent.TimeUnit;
 
 @Component
 public class EdcService {
 
-	private static final Logger logger = LoggerFactory.getLogger(MethodHandles.lookup().lookupClass());
-	private final HttpCallService httpCallService;
-	private final ObjectMapper objectMapper;
-	private final EdcProperties edcProperties;
+    private static final Logger logger = LoggerFactory.getLogger(MethodHandles.lookup().lookupClass());
+    private final HttpCallService httpCallService;
+    private final ObjectMapper objectMapper;
+    private final EdcProperties edcProperties;
 
-	public EdcService(HttpCallService httpCallService,
-					  ObjectMapper objectMapper,
-					  EdcProperties edcProperties) {
-		this.httpCallService = httpCallService;
-		this.objectMapper = objectMapper;
-		this.edcProperties = edcProperties;
-	}
+    public EdcService(HttpCallService httpCallService,
+                      ObjectMapper objectMapper,
+                      EdcProperties edcProperties) {
+        this.httpCallService = httpCallService;
+        this.objectMapper = objectMapper;
+        this.edcProperties = edcProperties;
+    }
 
 
-	/**
-	 * Rest call to get all contract offer and filter notification type contract
-	 */
-	public Optional<ContractOffer> findNotificationContractOffer(
-		String consumerEdcDataManagementUrl,
-		String providerConnectorControlPlaneIDSUrl,
-		Map<String, String> header,
-        boolean isInitialNotification
-	) throws IOException {
-		Catalog catalog = httpCallService.getCatalogFromProvider(consumerEdcDataManagementUrl, providerConnectorControlPlaneIDSUrl, header);
-		if (catalog.getContractOffers().isEmpty()) {
-			logger.error("No contract found");
-			throw new BadRequestException("Provider has no contract offers for us. Catalog is empty.");
-		}
-		logger.info(":::: Find Notification contract method[findNotificationContractOffer] total catalog ::{}", catalog.getContractOffers().size());
+    /**
+     * Rest call to get all contract offer and filter notification type contract
+     */
+    public Optional<ContractOffer> findNotificationContractOffer(
+            String consumerEdcDataManagementUrl,
+            String providerConnectorControlPlaneIDSUrl,
+            Map<String, String> header,
+            boolean isInitialNotification
+    ) throws IOException {
+        Catalog catalog = httpCallService.getCatalogFromProvider(consumerEdcDataManagementUrl, providerConnectorControlPlaneIDSUrl, header);
+        if (catalog.getContractOffers().isEmpty()) {
+            logger.error("No contract found");
+            throw new BadRequestException("Provider has no contract offers for us. Catalog is empty.");
+        }
+        logger.info(":::: Find Notification contract method[findNotificationContractOffer] total catalog ::{}", catalog.getContractOffers().size());
 
-        if (isInitialNotification){
+        if (isInitialNotification) {
             return catalog.getContractOffers().stream()
-                    .filter(it -> isPropertyQualityInvestigationType(it.getAsset()) && isPropertyReceiveNotificationMethod(it.getAsset()))
-                    .filter(this::hasTracePolicyConstraint)
+                    .filter(it -> it.getAsset().isQualityInvestigationReceive())
+                    .filter(this::hasTracePolicy)
                     .findAny();
         } else {
             return catalog.getContractOffers().stream()
-                    .filter(it -> isPropertyQualityInvestigationType(it.getAsset()) && isPropertyUpdateNotificationMethod(it.getAsset()))
-                    .filter(this::hasTracePolicyConstraint)
+                    .filter(it -> it.getAsset().isQualityInvestigationUpdate())
+                    .filter(this::hasTracePolicy)
                     .findAny();
         }
-	}
-
-    private boolean hasTracePolicyConstraint(ContractOffer contractOffer) {
-        Policy policy = contractOffer.getPolicy();
-
-        if (policy == null) {
-            return false;
-        }
-
-        List<Permission> permissions = policy.getPermissions();
-
-        if (permissions == null || permissions.isEmpty()) {
-            return false;
-        }
-
-        List<AtomicConstraint> atomicConstraints = permissions.stream()
-                .map(Rule::getConstraints)
-                .flatMap(Collection::stream)
-                .filter(it -> it instanceof AtomicConstraint)
-                .map(constraint -> (AtomicConstraint) constraint)
-                .toList();
-
-        if (atomicConstraints.isEmpty()) {
-            return false;
-        }
-
-        return atomicConstraints.stream().anyMatch(this::hasTracePurposeConstraint);
     }
 
-    private boolean hasTracePurposeConstraint(AtomicConstraint atomicConstraint) {
-        if (atomicConstraint == null) {
-            return false;
-        }
-
-        Expression leftExpression = atomicConstraint.getLeftExpression();
-        Expression rightExpression = atomicConstraint.getRightExpression();
-
-        if (leftExpression == null || rightExpression == null) {
-            return false;
-        }
-
-        if (!(leftExpression instanceof LiteralExpression) || !(rightExpression instanceof LiteralExpression)) {
-            return false;
-        }
-
-        LiteralExpression leftLiteralExpression = (LiteralExpression) atomicConstraint.getLeftExpression();
-        LiteralExpression rightLiteralExpression = (LiteralExpression) atomicConstraint.getRightExpression();
-
-        return matchesValue(leftLiteralExpression, "idsc:PURPOSE") && matchesValue(rightLiteralExpression, "ID 3.0 Trace");
+    private boolean hasTracePolicy(ContractOffer contractOffer) {
+        return contractOffer.getPolicy() != null && contractOffer.getPolicy().hasTracePolicy();
     }
 
-    private static boolean matchesValue(LiteralExpression literalExpression, String anObject) {
-        if (literalExpression.getValue() instanceof String literalExpressionValue) {
-            return literalExpressionValue.equals(anObject);
+    /**
+     * Prepare for contract negotiation. it will wait for while till API return agreementId
+     */
+    public String initializeContractNegotiation(String providerConnectorUrl, String assetId, String offerId, Policy policy, String consumerEdcUrl,
+                                                Map<String, String> header) throws InterruptedException, IOException {
+        // Initiate negotiation
+        ContractOfferDescription contractOfferDescription = new ContractOfferDescription(offerId, assetId, null, policy);
+        NegotiationInitiateRequestDto contractNegotiationRequest = NegotiationInitiateRequestDto.Builder.newInstance()
+                .offerId(contractOfferDescription).connectorId("provider").connectorAddress(providerConnectorUrl + edcProperties.getIdsPath())
+                .protocol("ids-multipart").build();
+
+        logger.info(":::: Start Contract Negotiation method[initializeContractNegotiation] offerId :{}, assetId:{}", offerId, assetId);
+
+        String negotiationId = initiateNegotiation(contractNegotiationRequest, consumerEdcUrl, header);
+        ContractNegotiationDto negotiation = null;
+
+        // Check negotiation state
+        while (negotiation == null || !negotiation.getState().equals("CONFIRMED")) {
+
+            logger.info(":::: waiting for contract to get confirmed");
+            ScheduledExecutorService scheduler = Executors.newScheduledThreadPool(1);
+            ScheduledFuture<ContractNegotiationDto> scheduledFuture =
+                    scheduler.schedule(() -> {
+                        var url = consumerEdcUrl + edcProperties.getNegotiationPath() + "/" + negotiationId;
+                        var request = new Request.Builder().url(url);
+                        header.forEach(request::addHeader);
+
+                        logger.info(":::: Start call for contract agreement method [initializeContractNegotiation] URL :{}", url);
+
+                        return (ContractNegotiationDto) httpCallService.sendRequest(request.build(), ContractNegotiationDto.class);
+                    }, 1000, TimeUnit.MILLISECONDS);
+            try {
+                negotiation = scheduledFuture.get();
+                scheduler.shutdown();
+            } catch (ExecutionException e) {
+                throw new RuntimeException(e);
+            } finally {
+                if (!scheduler.isShutdown()) {
+                    scheduler.shutdown();
+                }
+            }
         }
-
-        return false;
+        return negotiation.getContractAgreementId();
     }
 
-    private boolean isPropertyQualityInvestigationType(Asset asset) {
 
-		String formatted = String.format(":::: Asset %s has value %s", Constants.ASSET_KEY_NOTIFICATION_TYPE, asset.getPropertyNotificationType());
-		logger.info(formatted);
-		return Constants.ASSET_VALUE_QUALITY_INVESTIGATION.equals(asset.getPropertyNotificationType());
-	}
+    /**
+     * Rest call for Contract negotiation and return agreementId.
+     */
+    private String initiateNegotiation(NegotiationInitiateRequestDto contractOfferRequest, String consumerEdcDataManagementUrl,
+                                       Map<String, String> headers) throws IOException {
+        var url = consumerEdcDataManagementUrl + edcProperties.getNegotiationPath();
+        var requestBody = RequestBody.create(objectMapper.writeValueAsString(contractOfferRequest), Constants.JSON);
+        var request = new Request.Builder().url(url).post(requestBody);
 
-	private boolean isPropertyUpdateNotificationMethod(Asset asset) {
-		String formatted = String.format(":::: Asset isPropertyUpdateNotificationMethod %s has value %s", "asset:prop:notificationmethod", asset.getPropertyNotificationMethod());
-		logger.info(formatted);
-		return Constants.ASSET_VALUE_NOTIFICATION_METHOD_UPDATE.equals(asset.getPropertyNotificationMethod());
-	}
-
-    private boolean isPropertyReceiveNotificationMethod(Asset asset) {
-        String formatted = String.format(":::: Asset isPropertyReceiveNotificationMethod %s has value %s", "asset:prop:notificationmethod", asset.getPropertyNotificationMethod());
-        logger.info(formatted);
-        return Constants.ASSET_VALUE_NOTIFICATION_METHOD_RECEIVE.equals(asset.getPropertyNotificationMethod());
+        headers.forEach(request::addHeader);
+        request.build();
+        TransferId negotiationId = (TransferId) httpCallService.sendRequest(request.build(), TransferId.class);
+        logger.info(":::: Method [initiateNegotiation] Negotiation Id :{}", negotiationId.getId());
+        return negotiationId.getId();
     }
 
-	/**
-	 * Prepare for contract negotiation. it will wait for while till API return agreementId
-	 */
-	public String initializeContractNegotiation(String providerConnectorUrl, String assetId, String offerId, Policy policy, String consumerEdcUrl,
-												Map<String, String> header) throws InterruptedException, IOException {
-		// Initiate negotiation
-		ContractOfferDescription contractOfferDescription = new ContractOfferDescription(offerId, assetId, null, policy);
-		NegotiationInitiateRequestDto contractNegotiationRequest = NegotiationInitiateRequestDto.Builder.newInstance()
-			.offerId(contractOfferDescription).connectorId("provider").connectorAddress(providerConnectorUrl + edcProperties.getIdsPath())
-			.protocol("ids-multipart").build();
 
-		logger.info(":::: Start Contract Negotiation method[initializeContractNegotiation] offerId :{}, assetId:{}", offerId, assetId);
+    /**
+     * Rest call for Transfer Data with HttpProxy
+     */
+    public TransferId initiateHttpProxyTransferProcess(String agreementId, String assetId, String consumerEdcDataManagementUrl, String providerConnectorControlPlaneIDSUrl, Map<String, String> headers) throws IOException {
+        var url = consumerEdcDataManagementUrl + edcProperties.getTransferPath();
 
-		String negotiationId = initiateNegotiation(contractNegotiationRequest, consumerEdcUrl, header);
-		ContractNegotiationDto negotiation = null;
+        DataAddress dataDestination = DataAddress.Builder.newInstance().type("HttpProxy").build();
+        TransferType transferType = TransferType.Builder.transferType().contentType("application/octet-stream").isFinite(true).build();
 
-		// Check negotiation state
-		while (negotiation == null || !negotiation.getState().equals("CONFIRMED")) {
+        TransferRequestDto transferRequest = TransferRequestDto.Builder.newInstance()
+                .assetId(assetId).contractId(agreementId).connectorId("provider").connectorAddress(providerConnectorControlPlaneIDSUrl)
+                .protocol("ids-multipart").dataDestination(dataDestination).managedResources(false).transferType(transferType)
+                .build();
 
-			logger.info(":::: waiting for contract to get confirmed");
-			ScheduledExecutorService scheduler = Executors.newScheduledThreadPool(1);
-			ScheduledFuture<ContractNegotiationDto> scheduledFuture =
-				scheduler.schedule(() -> {
-					var url = consumerEdcUrl + edcProperties.getNegotiationPath() + "/" + negotiationId;
-					var request = new Request.Builder().url(url);
-					header.forEach(request::addHeader);
+        var requestBody = RequestBody.create(objectMapper.writeValueAsString(transferRequest), Constants.JSON);
+        var request = new Request.Builder().url(url).post(requestBody);
 
-					logger.info(":::: Start call for contract agreement method [initializeContractNegotiation] URL :{}", url);
-
-					return (ContractNegotiationDto) httpCallService.sendRequest(request.build(), ContractNegotiationDto.class);
-				}, 1000, TimeUnit.MILLISECONDS);
-			try {
-				negotiation = scheduledFuture.get();
-				scheduler.shutdown();
-			} catch (ExecutionException e) {
-				throw new RuntimeException(e);
-			} finally {
-				if (!scheduler.isShutdown()) {
-					scheduler.shutdown();
-				}
-			}
-		}
-		return negotiation.getContractAgreementId();
-	}
-
-
-	/**
-	 * Rest call for Contract negotiation and return agreementId.
-	 */
-	private String initiateNegotiation(NegotiationInitiateRequestDto contractOfferRequest, String consumerEdcDataManagementUrl,
-									   Map<String, String> headers) throws IOException {
-		var url = consumerEdcDataManagementUrl + edcProperties.getNegotiationPath();
-		var requestBody = RequestBody.create(objectMapper.writeValueAsString(contractOfferRequest), Constants.JSON);
-		var request = new Request.Builder().url(url).post(requestBody);
-
-		headers.forEach(request::addHeader);
-		request.build();
-		TransferId negotiationId = (TransferId) httpCallService.sendRequest(request.build(), TransferId.class);
-		logger.info(":::: Method [initiateNegotiation] Negotiation Id :{}", negotiationId.getId());
-		return negotiationId.getId();
-	}
-
-
-	/**
-	 * Rest call for Transfer Data with HttpProxy
-	 */
-	public TransferId initiateHttpProxyTransferProcess(String agreementId, String assetId, String consumerEdcDataManagementUrl, String providerConnectorControlPlaneIDSUrl, Map<String, String> headers) throws IOException {
-		var url = consumerEdcDataManagementUrl + edcProperties.getTransferPath();
-
-		DataAddress dataDestination = DataAddress.Builder.newInstance().type("HttpProxy").build();
-		TransferType transferType = TransferType.Builder.transferType().contentType("application/octet-stream").isFinite(true).build();
-
-		TransferRequestDto transferRequest = TransferRequestDto.Builder.newInstance()
-			.assetId(assetId).contractId(agreementId).connectorId("provider").connectorAddress(providerConnectorControlPlaneIDSUrl)
-			.protocol("ids-multipart").dataDestination(dataDestination).managedResources(false).transferType(transferType)
-			.build();
-
-		var requestBody = RequestBody.create(objectMapper.writeValueAsString(transferRequest), Constants.JSON);
-		var request = new Request.Builder().url(url).post(requestBody);
-
-		headers.forEach(request::addHeader);
-		logger.info(":::: call Transfer process with http Proxy method[initiateHttpProxyTransferProcess] agreementId:{} ,assetId :{},consumerEdcDataManagementUrl :{}, providerConnectorControlPlaneIDSUrl:{}", agreementId, assetId, consumerEdcDataManagementUrl, providerConnectorControlPlaneIDSUrl);
-		return (TransferId) httpCallService.sendRequest(request.build(), TransferId.class);
-	}
+        headers.forEach(request::addHeader);
+        logger.info(":::: call Transfer process with http Proxy method[initiateHttpProxyTransferProcess] agreementId:{} ,assetId :{},consumerEdcDataManagementUrl :{}, providerConnectorControlPlaneIDSUrl:{}", agreementId, assetId, consumerEdcDataManagementUrl, providerConnectorControlPlaneIDSUrl);
+        return (TransferId) httpCallService.sendRequest(request.build(), TransferId.class);
+    }
 }
